+++ main.go
@@ -5,6 +5,8 @@ package main
 import (
 	"github.com/goadesign/goa"
 	"github.com/goadesign/goa/middleware"
+	"github.com/spf13/afero"
+
 	"github.com/rchampourlier/letto_go/adapters"
 	"github.com/rchampourlier/letto_go/app"
 	"github.com/rchampourlier/letto_go/controllers"
@@ -15,6 +17,8 @@ func main() {
 	service := goa.New("letto")
 	s3 := adapters.NewS3("letto")
 
+	fs := afero.NewOsFs()
+
 	// Mount middleware
 	service.Use(middleware.RequestID())
 	service.Use(middleware.LogRequest(true))
@@ -23,7 +27,7 @@ func main() {
 
 	// Mount controllers
 	app.MountWorkflowController(service, controllers.NewWorkflowController(service, &s3))
-	app.MountTriggersController(service, controllers.NewTriggersController(service))
+	app.MountTriggersController(service, controllers.NewTriggersController(service, fs))
 
 	// Start service
 	if err := service.ListenAndServe(":9292"); err != nil {