+++ services/run_workflows.go
@@ -10,24 +10,28 @@ import (
 
 	"github.com/spf13/afero"
 
+	"github.com/rchampourlier/letto_go/events"
 	"github.com/rchampourlier/letto_go/exec"
 )
 
-var appFs = afero.NewOsFs()
 var tmpDirPrefix = "letto"
 
-// WorkflowData is a structure that is used to
-// pass data to execute the workflow with.
-type WorkflowData struct {
-	RequestBody    string
-	RequestHeaders map[string][]string
+// RunWorkflows stores the service's config.
+type RunWorkflows struct {
+	Fs afero.Fs
 }
 
-// RunWorkflows run all workflows for the specified `group`
-// with the specified data.
+// NewRunWorkflows creates a new RunWorkflowsService with
+// the specified `afero.Fs` filesystem abstraction.
+func NewRunWorkflows(fs afero.Fs) *RunWorkflows {
+	return &RunWorkflows{fs}
+}
+
+// OnReceivedWebhook runs all workflows for the specified
+// `group` with the specified event.
 //
 // TODO: refactor by extracting JS-related code to exec/js
-func RunWorkflows(group string, data WorkflowData) error {
+func (s *RunWorkflows) OnReceivedWebhook(event events.ReceivedWebhook) error {
 	rootDir, err := os.Getwd()
 	if err != nil {
 		return err
@@ -47,23 +51,23 @@ func RunWorkflows(group string, data WorkflowData) error {
 	// the `group`.
 	//group := ctx.Group
 	// TODO: selection by group
-	copyFile(path.Join(rootDir, "exec", "js", "main.js"), path.Join(tmpDir, "main.js"))
+	s.copyFile(path.Join(rootDir, "exec", "js", "main.js"), path.Join(tmpDir, "main.js"))
 
 	// Dump the request's data to a file passed to the script
-	requestJS, err := generateRequestJS(data)
+	requestJS, err := generateRequestJS(event.Body, event.Headers)
 	if err != nil {
 		return err
 	}
-	err = afero.WriteFile(appFs, path.Join(tmpDir, "request.js"), []byte(requestJS), 0777)
+	err = afero.WriteFile(s.Fs, path.Join(tmpDir, "request.js"), []byte(requestJS), 0777)
 	if err != nil {
 		return err
 	}
-	copyFile(path.Join(rootDir, "exec", "js", "data.js"), path.Join(tmpDir, "data.js"))
-	copyFile(path.Join(rootDir, "credentials.js"), path.Join(tmpDir, "credentials.js"))
+	s.copyFile(path.Join(rootDir, "exec", "js", "data.js"), path.Join(tmpDir, "data.js"))
+	s.copyFile(path.Join(rootDir, "credentials.js"), path.Join(tmpDir, "credentials.js"))
 
 	exec.RunJS(tmpDir, "./main.js")
 
-	err = appFs.RemoveAll(tmpDir)
+	err = s.Fs.RemoveAll(tmpDir)
 	if err != nil {
 		fmt.Printf("Could not remove tmp dir: %s\n", tmpDir)
 	}
@@ -72,9 +76,7 @@ func RunWorkflows(group string, data WorkflowData) error {
 }
 
 // TODO: handle requests where body is not JSON
-func generateRequestJS(data WorkflowData) (string, error) {
-	body := data.RequestBody
-	headers := data.RequestHeaders
+func generateRequestJS(body string, headers map[string][]string) (string, error) {
 	var requestJS string
 
 	if len(body) > 0 {
@@ -113,12 +115,12 @@ module.exports = {
 };
 `
 
-func copyFile(src string, dst string) error {
-	data, err := afero.ReadFile(appFs, src)
+func (s *RunWorkflows) copyFile(src string, dst string) error {
+	data, err := afero.ReadFile(s.Fs, src)
 	if err != nil {
 		return err
 	}
-	err = afero.WriteFile(appFs, dst, data, 0777)
+	err = afero.WriteFile(s.Fs, dst, data, 0777)
 	if err != nil {
 		return err
 	}